services/bin_service/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
PyMySQL==1.1.0
dj-database-url==2.1.0

# CORS
django-cors-headers==4.3.1

# MQTT Client
paho-mqtt==1.6.1

# Environment variables
python-decouple==3.8

# Utilities
requests==2.31.0
PyJWT==2.8.0
